Fix problem when lot of data
When there are more data than available pixels, the graph remains empty. This version fix the problem. See https://github.com/nnnick/Chart.js/issues/321

I'm  pretty sure that mathematically, it is not a valid solution  (for some data, the displayed graph could show something not reflecting the reality). But it is better than nothing...