options tickColor and minorTickColor
options tickColor and minorTickColor added; 
When the value of tickColor is not "gridLines" (which is the default value), major ticks drawn next to an axis and the axis have the same color. 
When the value of minorTickColor is not "gridLines" (which is the default value), minor ticks drawn next to an axis and the axis have the same color.